Drop Firefox privileges to (try to) play audio.
On a vaguely modern GNU/Linux distribution, PulseAudio is used to play sound on
the desktop. Given the TBB sets $HOME to its own directory, it doesn't find the
current user's existing instance of PulseAudio, and then tries to start its own.

Likely we don't want this to be allowed, in the context of these AppArmor
profiles:

* it's a pain to safely allow Firefox to run its own instance of PulseAudio;
* even if we managed to do it, it's not obvious to me that two concurrent
  isntances of PulseAudio, running as the same user, but now aware of the other,
  would work fine together in any useful way.

This can be revisited at a later point, but requires quite more work.