Comment out the "dbus" rule: it doesn't parse on any stable release of AppArmor yet.
One needs the upcoming AppArmor 2.9 userspace (or some Ubuntu's über-patched
snapshot of its development tree) to have this rule parse at all. And then, to
make it actually useful, one also needs to run a Linux kernel with patches that
have not been submitted to mainline yet.

In practice, with this change applied:

* Ubuntu users will need to un-comment this line, otherwise Firefox
  may disfunction.
* Basically all other users will be happy to see this profile actually loaded,
  and to see their TBB's Firefox be actually confined with AppArmor.

Once AppArmor 2.9 is available to most GNU/Linux users, this commit shall
be reverted.