Remove probably useless commented out AppArmor rules.
These rules are mostly duplicates from ones that are already in the `fonts`
abstraction, that we transitively include, so it's unclear to me when they
may be needed.

Note that the only difference between the abstraction's rules and these ones is
the "k" access granted on /var/cache/fontconfig/, which feels odd:  even if
a non-root user is allowed to do that, none of my other confined
fontconfig-using applications need that, so I would be surprised if Firefox used
fontconfig in a way that requires locking that directory. Still, if I'm shown
rejection logs that show it's needed, then I'm happy to add it back... and then,
perhaps it will make more sense to push it to the upstream fonts abstraction.