AppArmor: allow Firefox to learn how many CPU cores are present.
This was added in commits ace00d646 and a3908af8 already. Then, commit 04b24660
made the opposite change for reasons that are unknown to me.

Firefox can optimize things a bit depending on this information, which does not
seem terribly critical, and can probably be gathered by other means anyway,
given the current profile.

In the Firefox source code, this file is used in the PR_GetNumberOfProcessors
function, that itself is used e.g. in image/src/RasterImage.cpp to make use of
all available CPUs. Now, if someone shows that this information is leaked
on the network, and cannot be retrieved by other means given the same threat
model, then possibly it'll make sense to block this access... or to suggest
to the Tor Browser people to patch it out and always use one core only,
which would solve the (potential) problem also for people not using AppArmor.