AppArmor: allow Firefox to access mountinfo.
It's used in the GetDeviceName function (xpcom/io/nsLocalFileUnix.cpp), which is
in turn used by nsLocalFile::GetDiskSpaceAvailable. With this in mind, it's not
clear to me what the consequences of not letting Firefox access this information
are. E.g. it may very well let the user start downloading a file that won't fit
on disk.

So, unless good reasons are provided for blocking access to this file, I'm
allowing Firefox to read it.