AppArmor: grant access to mostly innocuous stuff Firefox tries to read.
I did not check in details why it needs that nowadays but this does not
increase the attack surface significantly, so let's allow it and don't
take the risk of breaking security critical stuff by denying it blindly.

If someone does the research and shows that it's safe to deny such access,
then we can do so.